Update estimator tests
