+++ qiskit_machine_learning/gradients/base/base_estimator_gradient.py
@@ -24,6 +24,7 @@ import numpy as np
 
 from qiskit.circuit import Parameter, ParameterExpression, QuantumCircuit
 from qiskit.primitives import BaseEstimator
+from qiskit.primitives.base import BaseEstimatorV2
 from qiskit.primitives.utils import _circuit_key
 from qiskit.providers import Options
 from qiskit.quantum_info.operators.base_operator import BaseOperator
@@ -46,7 +47,7 @@ class BaseEstimatorGradient(ABC):
 
     def __init__(
         self,
-        estimator: BaseEstimator,
+        estimator: BaseEstimator | BaseEstimatorV2,
         options: Options | None = None,
         derivative_type: DerivativeType = DerivativeType.REAL,
     ):
@@ -68,7 +69,7 @@ class BaseEstimatorGradient(ABC):
                 gradient and this type is the only supported type for function-level schemes like
                 finite difference.
         """
-        self._estimator: BaseEstimator = estimator
+        self._estimator = estimator
         self._default_options = Options()
         if options is not None:
             self._default_options.update_options(**options)
